requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
azure-ai-inference==1.0.0b9
azure-core==1.34.0
azure-storage-blob==12.25.1
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
contourpy==1.3.1
cryptography==45.0.4
curl_cffi==0.11.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.4.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docutils==0.21.2
einops==0.8.1
et_xmlfile==2.0.0
faiss-cpu==1.11.0
filelock==3.18.0
fonttools==4.55.8
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
grpcio==1.73.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.6
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
imagesize==1.4.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-huggingface==0.3.0
langchain-ollama==0.3.3
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.45
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.41.1
networkx==3.4.2
nomic==3.5.3
numpy==2.2.2
ollama==0.5.1
openai==1.86.0
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.1
pillow==10.4.0
platformdirs==4.3.8
portalocker==2.10.1
proglog==0.1.10
propcache==0.3.0
protobuf==6.31.1
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.9.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.1
pyparsing==3.2.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
streamlit==1.45.1
streamlit-chat==0.1.1
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.5.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
transformers==4.49.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
win32_setctime==1.2.0
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.62
zstandard==0.23.0
